+++ config/services.php
@@ -36,12 +36,6 @@ return [
         'redirect'      => env('GOOGLE_REDIRECT'),
     ],
 
-    'facebook' => [
-        'client_id'     => env('FACEBOOK_CLIENT_ID'),
-        'client_secret' => env('FACEBOOK_CLIENT_SECRET'),
-        'redirect'      => env('FACEBOOK_CALLBACK_URL'),
-    ],
-
     'webling' => [
         'api_url' => env('WEBLING_API_URL'),
         'api_key' => env('WEBLING_API_KEY'),